Dockerfile: 2-step enclave signing
Modified the Dockerfile to do 2-step signing. In the first step it
generates the enclave signing material file, where this file is supposed
to sign by an external signing tool (`enclave-config/sign-enclave.sh`).
The resulted signed hash file is passed to `sgx_sign` to generate the
final signed enclave.

The integration point for an external signing tool is the
`sign-enclave.sh` script defined under `enclave-config`. One has to
replace this with the suitable tool for signing while building the
docker image.