Makefile: new target for preparing the release branch
Added new make target named 'release-branch' which prepares a new branch.
It makes sure that all the deployment files point to the release image.

A new release branch is checked out from the current 'main'. The release
branch would be added with one additional commit with the
release-specific change like, modifying the deployment manifests to point
to the release image, and helm chart with the release number. The new
release branch is supposed to be tagged for the release.